requirements.txt
pinecone>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
markdown-it-py>=3.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1 